various
- Break up large source files
- improve UI consistency
- Add support for sending/receiving plaintext
- Detect when outbox plaintext is a URL and give clients extra context options
