+++ README.md
@@ -1,1 +1,6 @@
 # Axon
+
+## Color palette
+ - #202020
+ - #282828
+ - #303030
