+++ package.json
@@ -31,6 +31,7 @@
     "@electron/fuses": "1.8.0",
     "@vercel/webpack-asset-relocator-loader": "1.7.3",
     "babel-loader": "10.0.0",
+    "baseline-browser-mapping": "^2.9.19",
     "css-loader": "6.11.0",
     "electron": "38.1.2",
     "node-loader": "2.1.0",
