+++ src/app.jsx
@@ -1,6 +1,12 @@
 import React, { useEffect, useState, useRef } from "react";
 import { createRoot } from "react-dom/client";
-import QRCode from 'qrcode';
+
+
+import { handleAddText, handleChangedIP, openFile, handleDiscardPendingFile, initialize, updateURL } from "./handlers";
+import { Outbox, Inbox } from "./components/pages";
+import ShadedButton from "./components/ShadedButton";
+
+import ResponsiveButton from "./components/ResponsiveButton";
 
 // const root = createRoot(document.body);
 const root = createRoot(document.getElementById("root"));
@@ -39,6 +45,7 @@ function App() {
 
   const [activeSFile, setSelectedSFile] = useState(null);
   const [activeRFile, setSelectedRFile] = useState(null);
+  const [writingNewText, setWritingNewText] = useState(false);
 
   const [tlsKeyPath, setTLSKeyPath] = useState("");
   const [tlsCertPath, setTLSCertPath] = useState("");
@@ -46,6 +53,8 @@ function App() {
   const protocolRef = useRef(protocol);
   const ipRef = useRef(presentedIp);
   const portRef = useRef(port);
+
+  const [addTextValue, setAddTextValue] = useState("");
 
   useEffect(() => {
     protocolRef.current = protocol;
@@ -59,46 +68,79 @@ function App() {
     portRef.current = port;
   }, [port]);
 
+  useEffect(() => {
+    // In send mode:
+    //  when a file is selected, exit writing mode
+    if (activeSFile !== null) {
+      setWritingNewText(false);
+    }
+  }, [activeSFile]);
+
   //TODO :GET RID OF THIS!!!
   // REAL TODO: fix protocol, presentedIp, port being stale. (use react refs) 
-  function handleFileOpenClick(file = null) {
-    // console.log("stage 2 file=", file);
-    let _f = null;
-    if (file != null) {
-      _f = file;
-    }
-    // console.log("stage 3 file=", _f);
-    window.electronAPI.openSpecificFile(_f).then(([uid, fileName, fileSize]) => {
-      if (uid == "" || fileName == "null") {
-        return;
-      }
-      var url = `${protocolRef.current}://${ipRef.current}:${portRef.current}/get/${uid}`;
-      // console.log("appending to hostedfiles (current length " + hostedFiles.length + "): ", { id: uid, fileName: fileName, url: url, size: fileSize });
+  // function openFile(file = null) {
+  //   // console.log("stage 2 file=", file);
+  //   // let _f = null;
+  //   // if (file != null) {
+  //   //   _f = file;
+  //   // }
+  //   // console.log("stage 3 file=", _f);
+  //   window.electronAPI.openFile(file).then(([uid, fileName, fileSize]) => {
+  //     if (uid == "" || fileName == "null") {
+  //       return;
+  //     }
+  //     var url = `${protocolRef.current}://${ipRef.current}:${portRef.current}/get/${uid}`;
+  //     // console.log("appending to hostedfiles (current length " + hostedFiles.length + "): ", { id: uid, fileName: fileName, url: url, size: fileSize });
       
-      setHostedFiles(prev => {
-        const updated = [...prev, { id: uid, fileName: fileName, url: url, size: fileSize }]
-        setSelectedSFile(updated.length - 1);
-        return updated;
-      });
-    });
-  }
+  //     setHostedFiles(prev => {
+  //       const updated = [...prev, { id: uid, fileName: fileName, url: url, size: fileSize }]
+  //       setSelectedSFile(updated.length - 1);
+  //       return updated;
+  //     });
+  //   });
+  // }
 
-  function handleChangedIP(newIP) {
-    setPresentedIp(newIP);
-    // Save ip to config for next launch
-    window.electronAPI.setIP(newIP);
-  }
+  // function handleNewTextClick() {
+  //   setSelectedSFile(null);
+  //   setWritingNewText(true);
+  // }
 
-  function discardPendingFile() {
-    window.recvFileAPI.discardFile(pendingFiles[activeRFile].id);
-    var _pendingFiles = pendingFiles.filter(f => f.id !== pendingFiles[activeRFile].id);
-    if (_pendingFiles.length == 0) {
-      setSelectedRFile(null);
-    } else {
-      setSelectedRFile(0);
-    }
-    setPendingFiles(_pendingFiles);
-  }
+  // const handleAddText = (event) => {
+  //   event.preventDefault();
+  //   console.log("Received addtext event: ", event);
+
+  //   window.electronAPI.addTextToOutbox(addTextValue).then(([uid, filename, filesize]) => {
+  //     var url = `${protocolRef.current}://${ipRef.current}:${portRef.current}/get/${uid}`;
+
+  //     setHostedFiles(prev => {
+  //       const updated = [...prev, { id: uid, fileName: filename, url: url, size: filesize }]
+
+
+
+  //       setSelectedSFile(updated.length - 1);
+  //       return updated;
+  //     });
+  //   });
+
+  //   setAddTextValue("");
+  // }
+
+  // function handleChangedIP(newIP) {
+  //   setPresentedIp(newIP);
+  //   // Save ip to config for next launch
+  //   window.electronAPI.setIP(newIP);
+  // }
+
+  // function discardPendingFile() {
+  //   window.recvFileAPI.discardFile(pendingFiles[activeRFile].id);
+  //   var _pendingFiles = pendingFiles.filter(f => f.id !== pendingFiles[activeRFile].id);
+  //   if (_pendingFiles.length == 0) {
+  //     setSelectedRFile(null);
+  //   } else {
+  //     setSelectedRFile(0);
+  //   }
+  //   setPendingFiles(_pendingFiles);
+  // }
 
   //TODO GET RID OF THESE!!!
   // function sendMode() {
@@ -148,93 +190,15 @@ function App() {
 
   // initialization
   useEffect(() => {
+    console.log("renderer init: hostedFiles = \n" + hostedFiles.toString());
 
-    window.addEventListener("dragover", event => {
-      event.preventDefault();
-    });
-
-    window.addEventListener("drop", event => {
-      event.preventDefault();
-      const file = event.dataTransfer.files[0];
-      console.log("File dropped: ", file);
-      handleFileOpenClick(file);
-    });
-
-    async function getIP() {
-      // get default ip for dropdown
-      var ip = await window.electronAPI.getDefaultIP();
-      console.log("Got default IP: ", ip);
-      setPresentedIp(ip);
-      setRecvUrl(`${protocol}://${ip}:${port}/send`);
-    }
-    getIP();
-    async function getAddrs() {
-      // get all addresses for dropdown
-      let addrs = await window.electronAPI.listAddrs();
-      setAddrs(addrs);
-    }
-    getAddrs();
-
-    async function listenForNewUploads() {
-      window.recvFileAPI.onNewFile((file) => {
-        //Call this when a new file is received by the main process
-        setPendingFiles((prevPendingFiles) => [...prevPendingFiles, file]);
-      });
-    }
-    listenForNewUploads();
-
-    async function listenForSaveResults() {
-      window.recvFileAPI.onSaveFileResult((result) => {
-        console.log("Received save file result: ", result);
-        const parsed = result
-
-        const id = parsed.id;
-        setSavePaths((prev) => ({ ...prev, [id]: parsed.path }));
-        // console.log("SAVE PATH: " + parsed.path);
-        console.log(`Save result for file id ${id}: ${parsed.path}`);
-        if (parsed.path != null) {
-          // setPendingFiles((prevPendingFiles) => prevPendingFiles.filter((file) => file.id !== id));
-        }
-      });
-    }
-    listenForSaveResults();
-
-    window.electronAPI.getProtocol().then((savedProtocol) => {
-      setProtocol(savedProtocol);
-    });
-
-    async function getSavedSettings() {
-      let tlsKeyPath = await window.electronAPI.getTLSKeyPath();
-      setTLSKeyPath(tlsKeyPath);
-      let tlsCertPath = await window.electronAPI.getTLSCertPath();
-      setTLSCertPath(tlsCertPath);
-      let savedPort = await window.electronAPI.getPort();
-      setPort(savedPort);
-    }
-    getSavedSettings();
+    initialize(openFile, protocolRef, ipRef, portRef, setPort, setPresentedIp, setAddrs, setPendingFiles, setSavePaths, setProtocol, setTLSKeyPath, setTLSCertPath, setHostedFiles, setSelectedSFile);
   }, []);
 
   //update all URLs when port, protocol or presentedIp changes
   useEffect(() => {
-    // update recv url
-    setRecvUrl(`${protocol}://${presentedIp}:${port}/send`);
-
-    console.log(`Updated recvUrl to ${protocol}://${presentedIp}:${port}/send`);
-
-    if (port == "" || isNaN(port)) {
-      setPort(2222);
-    }
-
-    // also update all urls of hosted files
-    let newHostedFiles = hostedFiles.map((file) => {
-      let urlObj = new URL(file.url);
-      urlObj.protocol = protocol.toLowerCase();
-      urlObj.hostname = presentedIp;
-      urlObj.port = port.toString();
-      return { ...file, url: urlObj.toString() };
-    });
-    setHostedFiles(newHostedFiles);
-  }, [port, protocol, presentedIp]);
+    updateURL(protocol, presentedIp, port, setRecvUrl, hostedFiles, setHostedFiles);
+  }, [protocol, presentedIp, port]);
 
   // console.log("App rendering");
 
@@ -268,9 +232,9 @@ function App() {
   // }, [savePath]);
 
   function hasCurrentFileBeenSaved() {
-    let tmp = pendingFiles[activeRFile];
+    let current = pendingFiles[activeRFile];
     console.log("Active rfile:" + activeRFile);
-    return savePaths[tmp.id] != null;
+    return savePaths[current.id] != null;
   }
 
   const footerHeight = "24px";
@@ -301,11 +265,10 @@ function App() {
               <ResponsiveButton
                 key={btn.id}
                 label={btn.label}
-                buttonAction={() => {}}
+                buttonAction={() => setSelectedNavPage(i)}
                 selected={selectedNavPage === i}
-                setSelected={() => setSelectedNavPage(i)}
                 enabled={true}
-                customStyle={{display: "flex", width: "100%",  height: "45px", fontWeight: "regular", fontSize: "0.6rem", marginBottom: "3px",borderRadius: "4px", justifyContent: "center", alignItems: "center"}}
+                customStyle={{display: "flex", width: "100%",  height: "45px", fontWeight: "regular", fontSize: "0.6rem", marginBottom: "3px",borderRadius: "8px", justifyContent: "center", alignItems: "center"}}
                 shadeA={"#202020"}
                 shadeB={"#282828"}
                 shadeC={"#2c2c2c"}
@@ -318,170 +281,19 @@ function App() {
             label={"Settings"}
             buttonAction={() => setSelectedNavPage(-1)}
             selected={selectedNavPage === -1}
-            setSelected={() => setSelectedNavPage(-1)}
             enabled={true}
-            customStyle={{display: "flex", width: "100%",  height: "45px", fontWeight: "regular", fontSize: "0.6rem", marginBottom: "3px",borderRadius: "4px", justifyContent: "center", alignItems: "center"}}
+            customStyle={{display: "flex", width: "100%",  height: "45px", fontWeight: "regular", fontSize: "0.6rem", marginBottom: "3px", borderRadius: "8px", justifyContent: "center", alignItems: "center"}}
             shadeA={"#202020"}
             shadeB={"#282828"}
             shadeC={"#2c2c2c"}
           />
         </div>
-        <div className="content-wrapper" style={{height: "100%", width: "300px", flexGrow: 1}}>
-          {selectedNavPage == 0 ? (
-            <div style={{display: "flex", flexDirection: "row", height: "100%", width: "100%"}}>
-              <div id="left-summary-panel">
-                <div id="send-panel">
-                  <div className="left-send-header" style={{ display: "flex", justifyItems: "space-between", flexDirection: "column", marginBottom: "0", borderBottom: "1px solid #383838", paddingBottom: "4px" }}>
-                    <h4 style={{marginBottom: "5px", marginTop: "5px", marginLeft: "12px"}}>Outbox</h4>
-                    <div style={{display: "flex", flexDirection: "row", alignItems: "space-between", height: "25px"}}>
-                      <span className="simple-text" style={{margin: "auto", marginLeft: "13px", fontSize: "0.8rem", height: "fit-content"}}>{hostedFiles.length} file{hostedFiles.length !== 1 ? "s" : ""}</span>
-                      <div className="plus-btn" onClick={() => {handleFileOpenClick(null)}} />
-                    </div>
-                  </div>
-                  <div style={{marginLeft: "0", marginRight: "0"}}>
-                    {hostedFiles.map((file, i) => (
-
-                      <ResponsiveButton
-                        key={file.id}
-                        label={<SummaryListItem fileName={file.fileName.replace(/^.*[\\/]/, '')} onCloseClick={() => {var fileID = new URL(file.url).pathname.split("/").filter(Boolean).pop(); window.electronAPI.setServing(false, fileID); var _hostedFiles = hostedFiles.filter(f => f.id !== file.id); if (_hostedFiles.length == 0) {setSelectedSFile(null);} else {setSelectedSFile(0);} setHostedFiles(_hostedFiles);}} />}
-                        buttonAction={() => {}}
-                        selected={activeSFile === i}
-                        setSelected={() => setSelectedSFile(i)}
-                        enabled={true}
-                        customStyle={{}}
-                        shadeA={"#282828"}
-                        shadeB={"#303030"}
-                        shadeC={"#343434"}
-                      />
-                    ))}
-                  </div>
-                </div>
-              </div>
-              {activeSFile !== null ? (
-                <div id="right-detail-panel" style={{
-                verticalAlign: "top",
-                backgroundColor: "#303030",
-                borderRadius: "10px",
-                marginRight: "10px",
-                width: "200px",
-                flexGrow: 1,
-                marginTop: "10px",
-                marginBottom: "10px",
-                height: "calc(100% - 20px)",
-              }}>
-                <ServedItem key={hostedFiles[activeSFile]?.id} filename={hostedFiles[activeSFile]?.fileName} url={hostedFiles[activeSFile]?.url} size={hostedFiles[activeSFile]?.size} />
-              </div>
-              ) : null}
-            </div>
-          ) : selectedNavPage == 1 ? (
-            <div style={{display: "flex", flexDirection: "row", height: "100%", width: "100%"}}>
-              <div id="left-summary-panel">
-                <div id="recv-panel">
-                  <div className="left-recv-header" onClick={() => {setSelectedRFile(null);}} style={{ display: "flex", justifyItems: "space-between", flexDirection: "column", marginBottom: "0", borderBottom: "1px solid #383838", paddingBottom: "4px" }}>
-                    <h4 style={{marginBottom: "5px", marginTop: "5px", marginLeft: "12px" }}>Inbox</h4>
-                    <div style={{display: "flex", flexDirection: "row", alignItems: "space-between", height: "25px"}}>
-                      <span className="simple-text" style={{margin: "auto", marginLeft: "13px", fontSize: "0.8rem", height: "fit-content"}}>{pendingFiles.length} file{pendingFiles.length !== 1 ? "s" : ""}</span>
-                    </div>
-                  </div>
-
-                  <div style={{marginLeft: "0", marginRight: "0"}}>
-                    {pendingFiles.map((file, i) => (
-
-                      <ResponsiveButton
-                        key={file.id}
-                        label={<SummaryListItem fileName={file.filename.replace(/^.*[\\/]/, '')} onCloseClick={discardPendingFile} />}
-                        buttonAction={() => {}}
-                        selected={activeRFile === i}
-                        setSelected={() => setSelectedRFile(i)}
-                        enabled={true}
-                        customStyle={{}}
-                        shadeA={"#282828"}
-                        shadeB={"#303030"}
-                        shadeC={"#343434"}
-                      />
-                    ))}
-                  </div>
-                </div>
-              </div>
-              {activeRFile === null ? (
-                <div id="right-blank-panel" style={{width: "200px", flexGrow: 1}}>
-                  <div style={{color: "white", fontSize: "0.9rem", margin: "0 auto", width: "100%", textAlign: "center"}}>Share this QR code to allow others to send you files:</div>
-                  <div style={{width: "fit-content", margin: "0 auto", marginTop: "20px"}}>
-                    {recvUrl ? <QrComponent url={recvUrl} /> : <p>Loading QR...</p>}
-                  </div>
-                </div>
-              ) : (
-                <div id="right-detail-panel" style={{
-                  verticalAlign: "top",
-                  backgroundColor: "#303030",
-                  borderRadius: "10px",
-                  marginRight: "10px",
-                  width: "200px",
-                  flexGrow: 1,
-                  marginTop: "10px",
-                  marginBottom: "10px",
-                  height: "calc(100% - 20px)",
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "start",
-                  alignItems: "center",
-                  overflowWrap: "break-word",
-                  whiteSpace: "normal",
-                }}>
-                  <div style={{display: "flex", flexDirection: "row", width: "calc(100% - 10px)", marginRight: "10px"}}>
-                    <div style={{display: "flex", flexGrow: 1, margin: "10px", width: "calc(100% - 270px - 20px)"}}>
-                      <SimpleTextHeader primaryText={pendingFiles[activeRFile]?.filename} secondaryText={`Size: ${pendingFiles[activeRFile]?.size < 1024 ? `${pendingFiles[activeRFile]?.size} B` : pendingFiles[activeRFile]?.size < 1048576 ? `${(pendingFiles[activeRFile]?.size / 1024).toFixed(2)} KB` : `${(pendingFiles[activeRFile]?.size / 1048576).toFixed(2)} MB`}`} />
-                    </div>
-
-                    <div style={{width: "270px", display: "flex", flexDirection: "row", alignItems: "start", justifyContent: "space-around", fontSize: "13px"}}>
-
-                      <ResponsiveButton
-                        label={hasCurrentFileBeenSaved() ? "Saved" : "Save"}
-                        buttonAction={() => {window.recvFileAPI.saveFile(pendingFiles[activeRFile].id, ()=>{console.log("TESTING")});}}
-                        selected={false}
-                        setSelected={() => {}}
-                        enabled={!hasCurrentFileBeenSaved()}
-                        disabledStyle={{color: "#808080"}}
-                        customStyle={{display: "flex", width: "80px", height: "35px", borderRadius: "5px", justifyContent: "center", alignItems: "center", marginTop: "20px", marginLeft: "10px"}}
-                        shadeA={"#303030"}
-                        shadeB={"#383838"}
-                        shadeC={"#404040"}
-                      />
-
-                      <ResponsiveButton
-                        label={"Go to folder"}
-                        buttonAction={() => {window.recvFileAPI.revealFile(pendingFiles[activeRFile].id);}}
-                        selected={false}
-                        setSelected={() => {}}
-                        enabled={hasCurrentFileBeenSaved()}
-                        customStyle={{display: "flex", width: "80px", height: "35px", borderRadius: "5px", justifyContent: "center", alignItems: "center", marginTop: "20px", marginLeft: "10px"}}
-                        disabledStyle={{color: "#808080"}}
-                        shadeA={"#303030"}
-                        shadeB={"#383838"}
-                        shadeC={"#404040"}
-                      />
-
-                      <ResponsiveButton
-                        label={"Discard"}
-                        buttonAction={discardPendingFile}
-                        selected={false}
-                        setSelected={() => {}}
-                        enabled={true}
-                        customStyle={{display: "flex", width: "80px", height: "35px", borderRadius: "5px", justifyContent: "center", alignItems: "center", marginTop: "20px", marginLeft: "10px"}}
-                        shadeA={"#303030"}
-                        shadeB={"#983838"}
-                        shadeC={"#c04040"}
-                      />
-                    </div>
-                  </div>
-                  <div style={{display: "flex", flexDirection: "row", justifyContent: "center", alignItems: "flex-start", width: "100%", height: "40px", boxSizing: "border-box"}}>
-                    <div style={{width: "18px", flexGrow: 0}} />
-                    <div style={{width: "1px", flexGrow: 1, color: "#808080", fontStyle: "italic"}}>{savePaths[pendingFiles[activeRFile].id] ? "Saved to " + savePaths[pendingFiles[activeRFile].id] : ""}</div>
-                  </div>
-                </div>
-              )}
-            </div>
-          ) : selectedNavPage == -1 ? (
+        <div className="content-wrapper" style={{height: "100%", width: "300px", flexGrow: 1, borderRadius: "8px"}}>
+          {selectedNavPage === 0 ? (
+            <Outbox hostedFiles={hostedFiles} setHostedFiles={setHostedFiles} openFile={openFile} handleAddText={handleAddText} addTextValue={addTextValue} setAddTextValue={setAddTextValue} setSelectedSFile={setSelectedSFile} protocolRef={protocolRef} ipRef={ipRef} portRef={portRef} activeSFile={activeSFile}/>
+          ) : selectedNavPage === 1 ? (
+            <Inbox setSelectedRFile={setSelectedRFile} pendingFiles={pendingFiles} setPendingFiles={setPendingFiles} activeRFile={activeRFile} handleDiscardPendingFile={handleDiscardPendingFile} recvUrl={recvUrl} hasCurrentFileBeenSaved={hasCurrentFileBeenSaved} savePaths={savePaths} />
+          ) : selectedNavPage === -1 ? (
             <div style={{display: "flex", flexDirection: "column", height: "100%", width: "100%", fontSize: "0.8rem", marginLeft: "12px"}}>
               <h4 style={{marginBottom: "10px", marginTop: "13px"}}>Preferences</h4>
               <div style={{flexDirection: "column", display: "flex"}}>
@@ -503,7 +315,7 @@ function App() {
       <div style={{display: "flex", flexDirection: "row", justifyContent: "space-between", height: footerHeight, flexGrow: 0, width: "100%", backgroundColor: "#202020", lineHeight: "16px", color: "#606060", overflow: "hidden"}}>
         <div id="left-footer" style={{display: "flex", flexDirection: "row", alignItems: "center"}}>
           <div className="ip-selector" style={{margin: "0", width: "115px", height: footerHeight, overflow: "hidden"}}>
-            <select style={{height: "100%", fontSize: "12px", color: "#a0a0a0"}} onChange={(e) => handleChangedIP(e.target.value)} value={presentedIp}>
+            <select style={{height: "100%", fontSize: "12px", color: "#a0a0a0"}} onChange={(e) => handleChangedIP(e.target.value, setPresentedIp)} value={presentedIp}>
               {addrs.map((addr, index) => (
                 <option key={index} value={addr} style={{height: footerHeight}}>{addr}</option>
               ))}
@@ -554,7 +366,6 @@ function App() {
                 window.location.href = "https://github.com/sadam-04/Axon-node";
               }}
               selected={false}
-              setSelected={() => {}}
               enabled={true}
               customStyle={{height: footerHeight, fontSize: "12px", color: "#a0a0a0"}}
               shadeA={"#202020"}
@@ -568,38 +379,38 @@ function App() {
   );
 }
 
-function SummaryListItem({fileName, onCloseClick}) {
-  const [isHovered, setIsHovered] = useState(false);
+// function SummaryListItem({fileName, onCloseClick}) {
+//   const [isHovered, setIsHovered] = useState(false);
 
-  function handleMouseEnter(event) {
-    setIsHovered(true);
-  }
+//   function handleMouseEnter(event) {
+//     setIsHovered(true);
+//   }
 
-  function handleMouseLeave(event) {
-    setIsHovered(false);
-  }
+//   function handleMouseLeave(event) {
+//     setIsHovered(false);
+//   }
   
-  var _width = "260px";
+//   var _width = "260px";
 
-  return (
-    <div
-      onMouseEnter={(event) => {handleMouseEnter(event);}}
-      onMouseLeave={(event) => {handleMouseLeave(event);}}
-      style={{
-        padding: "6px 0 6px 10px",
-        display: "flex",
-        justifyContent: "space-between",
-        flexDirection: "row",
-        alignItems: "center",
-        width: `calc(${_width} - 20px)`,
-        height: "18px",
-      }}
-    >
-        <div style={{display: "block", fontSize: "13px", whiteSpace: "nowrap", overflow: "hidden", textOverflow: "ellipsis"}}>{fileName}</div>
-        {isHovered ? <div onClick={onCloseClick} className="outboxItemCloseBttn" style={{display: "block", width: "10px", height: "19px", marginRight: "8px"}}>✖</div> : null}
-    </div>
-  )
-}
+//   return (
+//     <div
+//       onMouseEnter={(event) => {handleMouseEnter(event);}}
+//       onMouseLeave={(event) => {handleMouseLeave(event);}}
+//       style={{
+//         padding: "6px 0 6px 10px",
+//         display: "flex",
+//         justifyContent: "space-between",
+//         flexDirection: "row",
+//         alignItems: "center",
+//         width: `calc(${_width} - 20px)`,
+//         height: "18px",
+//       }}
+//     >
+//         <div style={{display: "block", fontSize: "13px", whiteSpace: "nowrap", overflow: "hidden", textOverflow: "ellipsis"}}>{fileName}</div>
+//         {isHovered ? <div onClick={onCloseClick} className="outboxItemCloseBttn" style={{display: "block", width: "10px", height: "19px", marginRight: "8px"}}>✖</div> : null}
+//     </div>
+//   )
+// }
 
 // function ResponsiveButtonSet({buttonData}) {
 
@@ -610,139 +421,98 @@ function SummaryListItem({fileName, onCloseClick}) {
 //   }
 // }>
 
-function ResponsiveButton({selected, setSelected, enabled, buttonAction, onHover = null, label, customStyle = null, disabledStyle = null, shadeA, shadeB, shadeC}) {
-  const [clicked, setClicked] = useState(false);
-  const [hovered, setHovered] = useState(false);
-
-  return (
-  <div onMouseEnter={() => {setHovered(true); if (onHover) onHover();}} onMouseLeave={() => {setHovered(false); setClicked(false);}} onMouseDown={() => {setClicked(true)}} onMouseUp={() => {setSelected(); if (enabled) {buttonAction();} setClicked(false);}}>
-    <ShadedButton selected={selected} hovered={hovered} pressed={clicked} enabled={enabled} icon={label} customStyle={customStyle} disabledStyle={disabledStyle} shadeA={shadeA} shadeB={shadeB} shadeC={shadeC} />
-  </div>
-  );
-}
-
-function QrComponent({url}) {
-  const [src, setSrc] = useState("");
-  const [qrHoverMsg, setQrHoverMsg] = useState("Click to copy URL to clipboard");
-
-  // var combinedUrl = url ? url.replace("localhost", presentedHost ? presentedHost : "localhost") : "";
-
-  useEffect(() => {
-    (async () => {
-      try {
-        //var combinedUrl = url.replace("localhost", presentedHost ? presentedHost : "localhost");
-        console.log("Generating QR for url: " + url);
-        const dataUrl = await QRCode.toDataURL(url, {margin: 4});
-        setSrc(dataUrl);
-      } catch (err) {
-        console.error("Failed to generate QR code", err);
-      }
-    })();
-  }, [url]);
-
-  const handleQRClick = async (e) => {
-    window.focus();
-    navigator.clipboard.writeText(url).then(() => {
-      setQrHoverMsg("Copied!");
-    })
-  }
-
-  return (
-    <div className="qr-wrapper" onClick={handleQRClick} onMouseLeave={() => setQrHoverMsg("Click to copy URL to clipboard")}>
-      {src ? <img src={src} /> : <p>Loading QR...</p>}
-      <div className="qr-overlay"><div className="qr-overlay-text">{qrHoverMsg}</div></div>
-    </div>
-  );
-}
-
-function SimpleTextHeader({primaryText, postPrimaryContent=null, secondaryText}) {
-  return (
-    <div style={{width: "100%", textOverflow: "ellipsis", overflow: "hidden", whiteSpace: "nowrap"}}>
-      <strong>{primaryText}</strong>
-      {postPrimaryContent}
-      <br />
-      <ul style={{
-        listStyleType: "none",
-        paddingLeft: "8px",
-        marginTop: "4px",
-      }}>
-        <li>{secondaryText}</li>
-      </ul>
-    </div>
-  )
-}
-
-function ServedItem({filename, url, size}) {
-  // const [src, setSrc] = useState("");
-  const [checked, setChecked] = useState(true);
-  // const [qrHoverMsg, setQrHoverMsg] = useState("Click to copy URL to clipboard");
-  // const [fullUrl, setFullUrl] = useState("");
-
-  // let ip = "";
-  // const fullUrl = `http://${ip}:3030/get/${url}`;
-
-  const handleCheckboxChange = (e) => {
-    const isChecked = e.target.checked ? true : false;
-    setChecked(isChecked);
-
-    var fileID = new URL(url).pathname.split("/").filter(Boolean).pop()
-    window.electronAPI.setServing(isChecked, fileID);
-  };
-
-  // useEffect(() => {
-  //   (async () => {
-  //     try {
-  //       // ip = await window.electronAPI.getDefaultIP();
-  //       // console.log("renderer: set ip to " + ip);
-
-  //       // setFullUrl(`http://${ip}:3030/get/${url}`, async () => {
-  //         // console.log("renderer: fullUrl is " + fullUrl);
-  //       // setFullUrl(`http://${ip}:3030/get/${url}`);
 
 
-  //       // });
 
-  //     } catch (err) {
-  //       console.error("Failed to generate QR code", err);
-  //     }
-  //   })();
 
-  // }, [url]);
+// function SimpleTextHeader({primaryText, postPrimaryContent=null, secondaryText}) {
+//   return (
+//     <div style={{width: "100%", textOverflow: "ellipsis", overflow: "hidden", whiteSpace: "nowrap"}}>
+//       <strong>{primaryText}</strong>
+//       {postPrimaryContent}
+//       <br />
+//       <ul style={{
+//         listStyleType: "none",
+//         paddingLeft: "8px",
+//         marginTop: "4px",
+//       }}>
+//         <li>{secondaryText}</li>
+//       </ul>
+//     </div>
+//   )
+// }
 
-  filename = filename ? filename.replace(/^.*[\\/]/, '') : '';
+// function ServedItem({filename, url, size}) {
+//   // const [src, setSrc] = useState("");
+//   const [checked, setChecked] = useState(true);
+//   // const [qrHoverMsg, setQrHoverMsg] = useState("Click to copy URL to clipboard");
+//   // const [fullUrl, setFullUrl] = useState("");
 
-  const sizeString = size < 1024 ? `${size} B` : size < 1048576 ? `${(size / 1024).toFixed(2)} KB` : `${(size / 1048576).toFixed(2)} MB`;
+//   // let ip = "";
+//   // const fullUrl = `http://${ip}:3030/get/${url}`;
 
-  return <div className="file-entry">
+//   const handleCheckboxChange = (e) => {
+//     const isChecked = e.target.checked ? true : false;
+//     setChecked(isChecked);
 
-    <SimpleTextHeader primaryText={filename} postPrimaryContent={<input type="checkbox" checked={checked} onChange={handleCheckboxChange} />} secondaryText={`Size: ${sizeString}`} />
-    <div className="right-panel">
-      <QrComponent url={url} />
-    </div>
-  </div>
-};
+//     var fileID = new URL(url).pathname.split("/").filter(Boolean).pop()
+//     window.electronAPI.setServing(isChecked, fileID);
+//   };
+
+//   // useEffect(() => {
+//   //   (async () => {
+//   //     try {
+//   //       // ip = await window.electronAPI.getDefaultIP();
+//   //       // console.log("renderer: set ip to " + ip);
+
+//   //       // setFullUrl(`http://${ip}:3030/get/${url}`, async () => {
+//   //         // console.log("renderer: fullUrl is " + fullUrl);
+//   //       // setFullUrl(`http://${ip}:3030/get/${url}`);
+
+
+//   //       // });
+
+//   //     } catch (err) {
+//   //       console.error("Failed to generate QR code", err);
+//   //     }
+//   //   })();
+
+//   // }, [url]);
+
+//   filename = filename ? filename.replace(/^.*[\\/]/, '') : '';
+
+//   const sizeString = size < 1024 ? `${size} B` : size < 1048576 ? `${(size / 1024).toFixed(2)} KB` : `${(size / 1048576).toFixed(2)} MB`;
+
+//   return <div className="file-entry">
+
+//     <SimpleTextHeader primaryText={filename} postPrimaryContent={<input type="checkbox" checked={checked} onChange={handleCheckboxChange} />} secondaryText={`Size: ${sizeString}`} />
+//     <div className="right-panel">
+//       <QrComponent url={url} />
+//     </div>
+//   </div>
+// };
 
 // function Navbar() {
 
 // }
 
-function ShadedButton({ selected, hovered, pressed, enabled, customStyle = null, disabledStyle = null, icon, shadeA, shadeB, shadeC }) {
-  let shadeValue = shadeA;
+// function ShadedButton({ selected, hovered, pressed, enabled, customStyle = null, disabledStyle = null, icon, shadeA, shadeB, shadeC }) {
+//   let shadeValue = shadeA;
 
-  if (enabled) {
-    if ((selected && hovered && pressed) || (selected && !hovered && !pressed) || (!selected && hovered && !pressed)) {
-      shadeValue = shadeC;
-    } else if ((selected && hovered && !pressed) || (!selected && hovered && pressed)) {
-      shadeValue = shadeB;
-    }
-  }
+//   if (enabled) {
+//     if ((selected && hovered && pressed) || (selected && !hovered && !pressed) || (!selected && hovered && !pressed)) {
+//       shadeValue = shadeC;
+//     } else if ((selected && hovered && !pressed) || (!selected && hovered && pressed)) {
+//       shadeValue = shadeB;
+//     }
+//   }
 
-  return (
-    <div className="button-icon" style={{ ...customStyle, ...(enabled ? { backgroundColor: shadeValue } : disabledStyle) }}>
-      {icon}
-    </div>
-  );
-}
+//   return (
+//     <div className="button-icon" style={{ ...customStyle, ...(enabled ? { backgroundColor: shadeValue } : disabledStyle) }}>
+//       {icon}
+//     </div>
+//   );
+// }
 
 
 // root.render(<App />);
