+++ src/url-wrapper-2.html
@@ -1,0 +1,110 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+
+    <title>Axon</title>
+
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <style>
+      body {
+          font-family: Arial, sans-serif;
+          background-color: #101010;
+          color: #ccc;
+      }
+
+      .main-content {
+          background-color: #202020;
+
+          position: absolute;
+
+          /* max-width: 76%; */
+          width: 20%;
+          height: fit-content;
+
+          margin: 25px;
+          top: 0;
+          bottom: 0;
+          left: 0;
+          right: 0;
+
+          padding: 20px;
+
+          margin: auto;
+
+          /* box-shadow: inset 0 0 6px #252525; */
+          border-radius: 16px;
+      }
+
+      .section-title {
+          font-size: 20px;
+          margin-bottom: 15px;
+      }
+
+      #url-text {
+          word-break: break-all;
+          margin-bottom: 15px;
+      }
+
+      .button-wrapper {
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        width: 100%;
+        gap: 2px;
+        margin: auto;
+      }
+
+      .button {
+          padding: 10px;
+          font-size: 16px;
+          cursor: pointer;
+          /* background-color: #ccc; */
+          color: #ccc;
+          /* border: 1px solid #888; */
+          border-radius: 8px;
+          /* width: 11rem; */
+
+          transition: all 0.2s ease;
+      }
+
+      .button:hover {
+          background-color: #303030;
+          /* box-shadow: 0 0 6px #303030, 2px 2px 8px -4px #000000, inset 0 0 6px #363636; */
+          /* text-shadow: 0 0 8px #000; */
+      }
+
+      @media only screen and (max-width: 600px) {
+          .main-content {
+              width: 80%;
+          }
+      }
+    </style>
+
+    <script>
+      function open_link() {
+          window.open('{{url}}', '_blank');
+      }
+
+      function copy_link() {
+          navigator.clipboard.writeText('{{url}}');
+          document.querySelector('#copy-link').innerText = 'Copied!';
+      }
+    </script>
+
+  </head>
+  <body>
+    <div class="main-content">
+        <span class="section-title"><strong>URL</strong></span>
+
+        <div id="url-text">{{url}}</div>
+
+        <div style="color: #eee; font-weight: bold; margin: 30px 0 8px 0;">What do you want to do with this URL?</div>
+        <div class="button-wrapper" style="color: #eee;">
+            <div class="button" id="open-link" onClick="open_link()">Open in new tab</div>
+            <div class="button" id="copy-link" onClick="copy_link()" onmouseleave="this.innerText='Copy to clipboard'">Copy to clipboard</div>
+        </div>
+    </div>
+  </body>
+</html>
