+++ static/clientSend.css
@@ -6,19 +6,23 @@ body {
 	font-family:
     -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial,
     sans-serif;
+	display: flex;
+	justify-content: center;
+	align-items: center;
 }
 
-input {
+/* input {
 	font-weight: 400;
-}
+} */
 
 .main-content {
 	justify-content: center;
 	justify-items: center;
 	text-align: center;
-	width: 80%;
-	margin: auto;
-	display: block;
+	width: 600px;
+	display: flex;
+	flex-direction: column;
+	flex-grow: 1;
 }
 
 .section-title {
@@ -33,18 +37,19 @@ form {
 	flex-direction: column;
 	align-items: center;
 	padding: 20px;
+	border-radius: 8px;
 	/* border: 1px solid #ccc; */
-	width: 40%;
+	width: 300px;
 	margin: 20px auto;
 	/* box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); */
 	background-color: #282828;
 }
 
-@media only screen and (max-width: 768px) {
+/* @media only screen and (max-width: 768px) {
 	form {
 		width: 90%;
 	}
-}
+} */
 
 input[type="file"] {
 	display: none;
@@ -53,19 +58,20 @@ input[type="file"] {
 .form-field {
 	border: none;
 	outline: none;
-	padding: 10px 0px;
-	margin: 5px;
 	/* border: 1px solid #ccc; */
-	border-radius: 5px;
+	border-radius: 8px;
 	width: 85%;
 	/* transition: background-color 0.3s ease; */
-	font-size: 1em;
+	font-size: 0.8rem;
 	cursor: pointer;
+	box-sizing: border-box;
 }
 
 .bttn {
     background-color: #303030;
     color: #fff;
+	height: 30px;
+	padding: 0 10px;
 }
 
 .bttn:hover {
@@ -95,4 +101,33 @@ input[type="file"] {
 	100% {
 		transform: rotate(360deg);
 	}
+}
+
+.file-input {
+	border-radius: 8px;
+	height: 30px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding: 0;
+	margin: 0;
+}
+
+#file-name {
+	font-size: 0.8rem;
+	height: 34px;
+	line-height: 34px;
+	align-content: center;
+}
+
+#submit-button {
+	height: 30px;
+	border-radius: 8px;
+	font-weight: 400;
+	padding: 0 10px;
+	margin: 0;
+}
+
+#text {
+	margin-bottom: 6px;
 }
