+++ webpack.renderer.config.js
@@ -5,6 +5,11 @@ rules.push({
   use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
 });
 
+rules.push({
+  test: /\.(png|jpe?g|gif|svg)$/i,
+  type: "asset/resource",
+});
+
 module.exports = {
   // Put your normal webpack config below here
   module: {
